Fuel for the code fire: don't prefix map with the explicit import.
Map is imported implicitly as part of the Prelude module. Referring to
it via the explicit import qualifier of the Data.List module is thus,
 from a pragmatic perspective, redundant.

It can be argued, however, that the explicit module reference adds
clarity and prevents future breakage should map suddenly be dropped
from Prelude.

I've gravitated towards the second view in the past. In recent times,
however, I've become more inclined to save the extra keystrokes and make
use of the implicit import.

Let's see how this experiment goes.